Dockerfile|FindGMock.cmake: added GMock module to cmake
We use Google Test and Google Mock in more than a couple places, it
makes sense to centralize it. This should really be a standard CMake
module, but until then we can supply our own.

ARG variables are local to Dockerfile, so this does not clutter run-time
environment

This module was pulled from our local sources, but is originally
from...?